requirements.txt
annotated-types==0.7.0
certifi==2024.7.4
fastapi==0.111.1
Jinja2==3.1.4
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
SQLAlchemy==2.0.31
starlette==0.37.2
uvicorn==0.30.3
